added capture dir

+++ src/main/kotlin/com/api/characters/service/CharacterService.kt
@@ -5,6 +5,7 @@ import com.api.characters.repository.CharacterRepository
 import org.springframework.core.io.Resource
 import org.springframework.core.io.UrlResource
 import org.springframework.stereotype.Service
+import org.springframework.transaction.annotation.Transactional
 import org.springframework.web.multipart.MultipartFile
 import java.io.IOException
 import java.net.InetAddress
@@ -16,36 +17,37 @@ class CharacterService(
     private val characterRepository:CharacterRepository
 ):ICharacterService{
 
-    private val serverIpAddress = InetAddress.getLocalHost().hostAddress
     private val uploadDirectory = "uploads/images"
 
-    override fun save(character: Character, image:MultipartFile): Character {
-
+    @Transactional(rollbackFor = [IOException::class])
+    override fun save(character: Character, image: MultipartFile): Character {
         val imagesDir = Paths.get(uploadDirectory)
 
         if (!Files.exists(imagesDir)) {
-            try {
-                Files.createDirectories(imagesDir)
-            } catch (e: IOException) {
-                e.printStackTrace()
-                throw Exception("Failed to create upload directory!")
-            }
+            Files.createDirectories(imagesDir)
         }
 
-        val fileName = character.name?.replace("\\s".toRegex(), "_")?.toLowerCase()
-        val filePath = imagesDir.resolve("$fileName.jpg")
+        val fileName = character.name?.replace("\\s".toRegex(), "_")?.toLowerCase() + ".jpg"
+        val filePath = imagesDir.resolve(fileName)
+
+        if (Files.exists(filePath)) {
+            throw Exception("Image with the same name already exists!")
+        }
 
         try {
             Files.copy(image.inputStream, filePath)
         } catch (e: IOException) {
-            e.printStackTrace()
-            throw Exception("failed to save image!")
+            throw IOException("Failed to save image", e)
         }
 
-        val imageLink = "http://localhost:8080/api/character/image/$fileName.jpg"
+        val imageLink = "http://localhost:8080/api/character/image/$fileName"
         character.image = imageLink
 
-        return characterRepository.save(character)
+        try {
+            return characterRepository.save(character)
+        } catch (e: Exception) {
+            throw RuntimeException("Failed to save character", e)
+        }
     }
     fun loadImage(imageName: String): Resource {
         try {
